Update auth inheritance hierarchy
